requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
